Refactor model picker module
